requirements.txt
awesome-slugify==1.6.5
certifi==2017.11.5
chardet==3.0.4
click==6.7
idna==2.6
regex==2017.11.9
requests==2.18.4
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
xdg==2.0.0
